// apps/trade-pro/src/pages/Taxation/ItemTaxShedule/type.ts
export type TItemTaxShedule = {
  Id: number;
  IsActive: boolean;
  EffectedDate: Date;
  EntryDate: Date;
  ModifyDate: Date;
  CompanyId: number;
  EntryUser: number;
  ItemId: number;
  ModifyUser: number;
  OrganizationId: number;
  TaxTypeId: number;
  TaxName: string;
  ItemName: string;
};
// ya nabi sb karam ha tumhara natt